+++ trunk/src/modp_b64.h
@@ -31,6 +31,7 @@
 #endif
 
 BEGIN_C
+#include "modp_stdint.h"
 
 /**
  * Encode a raw binary string into base 64.
@@ -57,7 +58,7 @@ BEGIN_C
  * \endcode
  *
  */
-int modp_b64_encode(char* dest, const char* str, int len);
+size_t modp_b64_encode(char* dest, const char* str, size_t len);
 
 /**
  * Decode a base64 encoded string
@@ -81,7 +82,7 @@ int modp_b64_encode(char* dest, const char* str, int len);
  * if (len == -1) { error }
  * \endcode
  */
-int modp_b64_decode(char* dest, const char* src, int len);
+size_t modp_b64_decode(char* dest, const char* src, size_t len);
 
 /**
  * Given a source string of length len, this returns the amount of
